+++ examples/lwIP-SAM-D21-Curiosity-Nano/firmware/src/main.c
@@ -60,9 +60,11 @@ Microchip or any third party.
 /*                          USER ADJUSTABLE                             */
 /*>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>*/
 
-#define FIRMWARE_VERSION            "V3.1.0"
+#define FIRMWARE_VERSION            "V3.1.1"
 
+#ifndef BOARD_INSTANCE
 #define BOARD_INSTANCE              (0)
+#endif
 #define T1S_PLCA_ENABLE             (true)
 #define T1S_PLCA_NODE_ID            (BOARD_INSTANCE + 1)
 #define T1S_PLCA_NODE_COUNT         (8)