+++ examples/lwIP-SAM-D21-Curiosity-Nano/firmware/src/tc6-stub.c
@@ -81,7 +81,7 @@ typedef struct
     uint8_t intReported;
     uint8_t idx;
     bool opened;
-    bool busy;
+    volatile bool busy;
     volatile uint8_t macValid;
 } Stub_Local_t;