+++ examples/noIP-PIC18-Curiosity-Nano/PIC18.X/mcc_generated_files/mcc.h
@@ -13,12 +13,12 @@
   @Description:
     This header file provides implementations for driver APIs for all modules selected in the GUI.
     Generation Information :
-        Product Revision  :  PIC10 / PIC12 / PIC16 / PIC18 MCUs - 1.81.7
+        Product Revision  :  PIC10 / PIC12 / PIC16 / PIC18 MCUs - 1.81.8
         Device            :  PIC18F57Q43
         Driver Version    :  2.00
     The generated drivers are tested against the following:
-        Compiler          :  XC8 2.31 and above or later
-        MPLAB             :  MPLAB X 5.45
+        Compiler          :  XC8 2.36 and above or later
+        MPLAB             :  MPLAB X 6.00
 */
 
 /*
@@ -53,11 +53,12 @@
 #include <stdbool.h>
 #include <conio.h>
 #include "interrupt_manager.h"
-#include "dma1.h"
-#include "ext_int.h"
-#include "dma2.h"
+#include "i2c1_master.h"
+#include "tmr2.h"
 #include "tmr0.h"
+#include "ext_int.h"
 #include "uart1.h"
+#include "drivers/i2c_simple_master.h"
 #include "spi1.h"
 
 
@@ -101,17 +102,6 @@ void OSCILLATOR_Initialize(void);
  */
 void PMD_Initialize(void);
 
-/**
- * @Param
-    none
- * @Returns
-    none
- * @Description
-    Initializes the System Arbiter for DMA to the default priority.
- * @Example
-    SystemArbiter_DMA_Initialize();
- */
-void SystemArbiter_Initialize(void);
 
 #endif	/* MCC_H */
 /**