+++ examples/noIP-PIC18-Curiosity-Nano/PIC18.X/mcc_generated_files/pin_manager.c
@@ -13,36 +13,36 @@
   Description:
     This header file provides implementations for pin APIs for all pins selected in the GUI.
     Generation Information :
-        Product Revision  :  PIC10 / PIC12 / PIC16 / PIC18 MCUs - 1.81.7
+        Product Revision  :  PIC10 / PIC12 / PIC16 / PIC18 MCUs - 1.81.8
         Device            :  PIC18F57Q43
         Driver Version    :  2.11
     The generated drivers are tested against the following:
-        Compiler          :  XC8 2.31 and above
-        MPLAB             :  MPLAB X 5.45
+        Compiler          :  XC8 2.36 and above
+        MPLAB             :  MPLAB X 6.00
 
     Copyright (c) 2013 - 2015 released Microchip Technology Inc.  All rights reserved.
 */
 
 /*
-    (c) 2018 Microchip Technology Inc. and its subsidiaries. 
-    
-    Subject to your compliance with these terms, you may use Microchip software and any 
-    derivatives exclusively with Microchip products. It is your responsibility to comply with third party 
-    license terms applicable to your use of third party software (including open source software) that 
+    (c) 2018 Microchip Technology Inc. and its subsidiaries.
+
+    Subject to your compliance with these terms, you may use Microchip software and any
+    derivatives exclusively with Microchip products. It is your responsibility to comply with third party
+    license terms applicable to your use of third party software (including open source software) that
     may accompany Microchip software.
-    
-    THIS SOFTWARE IS SUPPLIED BY MICROCHIP "AS IS". NO WARRANTIES, WHETHER 
-    EXPRESS, IMPLIED OR STATUTORY, APPLY TO THIS SOFTWARE, INCLUDING ANY 
-    IMPLIED WARRANTIES OF NON-INFRINGEMENT, MERCHANTABILITY, AND FITNESS 
+
+    THIS SOFTWARE IS SUPPLIED BY MICROCHIP "AS IS". NO WARRANTIES, WHETHER
+    EXPRESS, IMPLIED OR STATUTORY, APPLY TO THIS SOFTWARE, INCLUDING ANY
+    IMPLIED WARRANTIES OF NON-INFRINGEMENT, MERCHANTABILITY, AND FITNESS
     FOR A PARTICULAR PURPOSE.
-    
-    IN NO EVENT WILL MICROCHIP BE LIABLE FOR ANY INDIRECT, SPECIAL, PUNITIVE, 
-    INCIDENTAL OR CONSEQUENTIAL LOSS, DAMAGE, COST OR EXPENSE OF ANY KIND 
-    WHATSOEVER RELATED TO THE SOFTWARE, HOWEVER CAUSED, EVEN IF MICROCHIP 
-    HAS BEEN ADVISED OF THE POSSIBILITY OR THE DAMAGES ARE FORESEEABLE. TO 
-    THE FULLEST EXTENT ALLOWED BY LAW, MICROCHIP'S TOTAL LIABILITY ON ALL 
-    CLAIMS IN ANY WAY RELATED TO THIS SOFTWARE WILL NOT EXCEED THE AMOUNT 
-    OF FEES, IF ANY, THAT YOU HAVE PAID DIRECTLY TO MICROCHIP FOR THIS 
+
+    IN NO EVENT WILL MICROCHIP BE LIABLE FOR ANY INDIRECT, SPECIAL, PUNITIVE,
+    INCIDENTAL OR CONSEQUENTIAL LOSS, DAMAGE, COST OR EXPENSE OF ANY KIND
+    WHATSOEVER RELATED TO THE SOFTWARE, HOWEVER CAUSED, EVEN IF MICROCHIP
+    HAS BEEN ADVISED OF THE POSSIBILITY OR THE DAMAGES ARE FORESEEABLE. TO
+    THE FULLEST EXTENT ALLOWED BY LAW, MICROCHIP'S TOTAL LIABILITY ON ALL
+    CLAIMS IN ANY WAY RELATED TO THIS SOFTWARE WILL NOT EXCEED THE AMOUNT
+    OF FEES, IF ANY, THAT YOU HAVE PAID DIRECTLY TO MICROCHIP FOR THIS
     SOFTWARE.
 */
 
@@ -51,7 +51,7 @@
 
 
 
-void (*IOCAF6_InterruptHandler)(void);
+void (*IOCBF5_InterruptHandler)(void);
 
 
 void PIN_MANAGER_Initialize(void)
@@ -60,7 +60,7 @@ void PIN_MANAGER_Initialize(void)
     LATx registers
     */
     LATE = 0x00;
-    LATD = 0x10;
+    LATD = 0xD0;
     LATA = 0x00;
     LATF = 0x00;
     LATB = 0x00;
@@ -71,20 +71,20 @@ void PIN_MANAGER_Initialize(void)
     */
     TRISE = 0x07;
     TRISF = 0xF6;
-    TRISA = 0x79;
-    TRISB = 0xFF;
+    TRISA = 0xFF;
+    TRISB = 0xF9;
     TRISC = 0xAF;
-    TRISD = 0xEF;
+    TRISD = 0x0F;
 
     /**
     ANSELx registers
     */
-    ANSELD = 0xEF;
-    ANSELC = 0x8F;
-    ANSELB = 0xEF;
+    ANSELD = 0x0F;
+    ANSELC = 0x87;
+    ANSELB = 0xC9;
     ANSELE = 0x07;
     ANSELF = 0xF4;
-    ANSELA = 0x31;
+    ANSELA = 0xBF;
 
     /**
     WPUx registers
@@ -110,7 +110,7 @@ void PIN_MANAGER_Initialize(void)
     ODCONE = 0x00;
     ODCONF = 0x00;
     ODCONA = 0x00;
-    ODCONB = 0x00;
+    ODCONB = 0x06;
     ODCONC = 0x00;
     ODCOND = 0x00;
 
@@ -119,7 +119,7 @@ void PIN_MANAGER_Initialize(void)
     */
     SLRCONA = 0xFF;
     SLRCONB = 0xFF;
-    SLRCONC = 0xC7;
+    SLRCONC = 0xFF;
     SLRCOND = 0xFF;
     SLRCONE = 0x07;
     SLRCONF = 0xFF;
@@ -136,68 +136,74 @@ void PIN_MANAGER_Initialize(void)
 
 
     /**
-    IOCx registers 
+    IOCx registers
     */
-    //interrupt on change for group IOCAF - flag
-    IOCAFbits.IOCAF6 = 0;
-    //interrupt on change for group IOCAN - negative
-    IOCANbits.IOCAN6 = 1;
-    //interrupt on change for group IOCAP - positive
-    IOCAPbits.IOCAP6 = 1;
+    //interrupt on change for group IOCBF - flag
+    IOCBFbits.IOCBF5 = 0;
+    //interrupt on change for group IOCBN - negative
+    IOCBNbits.IOCBN5 = 1;
+    //interrupt on change for group IOCBP - positive
+    IOCBPbits.IOCBP5 = 0;
 
 
 
     // register default IOC callback functions at runtime; use these methods to register a custom function
-    IOCAF6_SetInterruptHandler(IOCAF6_DefaultInterruptHandler);
-   
-    
-	
-    SPI1SCKPPS = 0x16;   //RC6->SPI1:SCK1;    
-    INT0PPS = 0x06;   //RA6->EXT_INT:INT0;    
-    RC4PPS = 0x32;   //RC4->SPI1:SDO1;    
-    RC6PPS = 0x31;   //RC6->SPI1:SCK1;    
-    RF0PPS = 0x20;   //RF0->UART1:TX1;    
-    U1RXPPS = 0x29;   //RF1->UART1:RX1;    
-    SPI1SDIPPS = 0x15;   //RC5->SPI1:SDI1;    
+    IOCBF5_SetInterruptHandler(IOCBF5_DefaultInterruptHandler);
+
+    // Enable IOCI interrupt
+    PIE0bits.IOCIE = 1;
+
+
+    SPI1SCKPPS = 0x16;   //RC6->SPI1:SCK1;
+    I2C1SDAPPS = 0x0A;   //RB2->I2C1:SDA1;
+    RB1PPS = 0x37;   //RB1->I2C1:SCL1;
+    INT0PPS = 0x0D;   //RB5->EXT_INT:INT0;
+    RB2PPS = 0x38;   //RB2->I2C1:SDA1;
+    RC4PPS = 0x32;   //RC4->SPI1:SDO1;
+    RC6PPS = 0x31;   //RC6->SPI1:SCK1;
+    RF0PPS = 0x20;   //RF0->UART1:TX1;
+    U1RXPPS = 0x29;   //RF1->UART1:RX1;
+    I2C1SCLPPS = 0x09;   //RB1->I2C1:SCL1;
+    SPI1SDIPPS = 0x15;   //RC5->SPI1:SDI1;
 }
-  
+
 void PIN_MANAGER_IOC(void)
-{   
-	// interrupt on change for pin IOCAF6
-    if(IOCAFbits.IOCAF6 == 1)
+{
+	// interrupt on change for pin IOCBF5
+    if(IOCBFbits.IOCBF5 == 1)
     {
-        IOCAF6_ISR();  
-    }	
+        IOCBF5_ISR();
+    }
 }
 
 /**
-   IOCAF6 Interrupt Service Routine
+   IOCBF5 Interrupt Service Routine
 */
-void IOCAF6_ISR(void) {
+void IOCBF5_ISR(void) {
 
-    // Add custom IOCAF6 code
+    // Add custom IOCBF5 code
 
     // Call the interrupt handler for the callback registered at runtime
-    if(IOCAF6_InterruptHandler)
+    if(IOCBF5_InterruptHandler)
     {
-        IOCAF6_InterruptHandler();
+        IOCBF5_InterruptHandler();
     }
-    IOCAFbits.IOCAF6 = 0;
+    IOCBFbits.IOCBF5 = 0;
 }
 
 /**
-  Allows selecting an interrupt handler for IOCAF6 at application runtime
+  Allows selecting an interrupt handler for IOCBF5 at application runtime
 */
-void IOCAF6_SetInterruptHandler(void (* InterruptHandler)(void)){
-    IOCAF6_InterruptHandler = InterruptHandler;
+void IOCBF5_SetInterruptHandler(void (* InterruptHandler)(void)){
+    IOCBF5_InterruptHandler = InterruptHandler;
 }
 
 /**
-  Default interrupt handler for IOCAF6
+  Default interrupt handler for IOCBF5
 */
-void IOCAF6_DefaultInterruptHandler(void){
-    // add your IOCAF6 interrupt custom code
-    // or set custom function using IOCAF6_SetInterruptHandler()
+void IOCBF5_DefaultInterruptHandler(void){
+    // add your IOCBF5 interrupt custom code
+    // or set custom function using IOCBF5_SetInterruptHandler()
 }
 
 /**