+++ examples/noIP-PIC18-Curiosity-Nano/PIC18.X/mcc_generated_files/spi1.h
@@ -13,12 +13,12 @@
   @Description
     This header file provides APIs for driver for SPI1.
     Generation Information :
-        Product Revision  :  PIC10 / PIC12 / PIC16 / PIC18 MCUs - 1.81.7
+        Product Revision  :  PIC10 / PIC12 / PIC16 / PIC18 MCUs - 1.81.8
         Device            :  PIC18F57Q43
-        Driver Version    :  3.0.0
+        Driver Version    :  1.0.0
     The generated drivers are tested against the following:
-        Compiler          :  XC8 2.31 and above or later
-        MPLAB             :  MPLAB X 5.45
+        Compiler          :  XC8 2.36 and above or later
+        MPLAB             :  MPLAB X 6.00
 */
 
 /*
@@ -64,11 +64,11 @@ void SPI1_Initialize(void);
 bool SPI1_Open(spi1_modes_t spi1UniqueConfiguration);
 void SPI1_Close(void);
 uint8_t SPI1_ExchangeByte(uint8_t data);
-void SPI1_ExchangeBlocks(const uint8_t *pTx, uint8_t *pRx, size_t blockSize);
 void SPI1_ExchangeBlock(void *block, size_t blockSize);
 void SPI1_WriteBlock(void *block, size_t blockSize);
 void SPI1_ReadBlock(void *block, size_t blockSize);
 void SPI1_WriteByte(uint8_t byte);
 uint8_t SPI1_ReadByte(void);
+void SPI1_ExchangeBlocks(const uint8_t *pTx, uint8_t *pRx, size_t blockSize);
 
 #endif //SPI1_H