+++ examples/noIP-PIC18-Curiosity-Nano/PIC18.X/mcc_generated_files/uart1.h
@@ -13,12 +13,12 @@
   @Description
     This header file provides APIs for driver for UART1.
     Generation Information :
-        Product Revision  :  PIC10 / PIC12 / PIC16 / PIC18 MCUs - 1.81.7
+        Product Revision  :  PIC10 / PIC12 / PIC16 / PIC18 MCUs - 1.81.8
         Device            :  PIC18F57Q43
         Driver Version    :  2.4.1
     The generated drivers are tested against the following:
-        Compiler          :  XC8 2.31 and above
-        MPLAB             :  MPLAB X 5.45
+        Compiler          :  XC8 2.36 and above
+        MPLAB             :  MPLAB X 6.00
 */
 
 /*
@@ -82,11 +82,6 @@ typedef union {
     uint8_t status;
 }uart1_status_t;
 
-/**
- Section: Global variables
- */
-extern volatile uint8_t uart1TxBufferRemaining;
-extern volatile uint8_t uart1RxCount;
 
 /**
   Section: UART1 APIs
@@ -380,68 +375,7 @@ uint8_t UART1_Read(void);
 */
 void UART1_Write(uint8_t txData);
 
-/**
-  @Summary
-    Maintains the driver's transmitter state machine and implements its ISR.
 
-  @Description
-    This routine is used to maintain the driver's internal transmitter state
-    machine.This interrupt service routine is called when the state of the
-    transmitter needs to be maintained in a non polled manner.
-
-  @Preconditions
-    UART1_Initialize() function should have been called
-    for the ISR to execute correctly.
-
-  @Param
-    None
-
-  @Returns
-    None
-*/     
-void UART1_Transmit_ISR(void);
-
-/**
-  @Summary
-    Maintains the driver's receiver state machine and implements its ISR
-
-  @Description
-    This routine is used to maintain the driver's internal receiver state
-    machine.This interrupt service routine is called when the state of the
-    receiver needs to be maintained in a non polled manner.
-
-  @Preconditions
-    UART1_Initialize() function should have been called
-    for the ISR to execute correctly.
-
-  @Param
-    None
-
-  @Returns
-    None
-*/       
-void UART1_Receive_ISR(void);
-
-/**
-  @Summary
-    Maintains the driver's receiver state machine
-
-  @Description
-    This routine is called by the receive state routine and is used to maintain 
-    the driver's internal receiver state machine. It should be called by a custom
-    ISR to maintain normal behavior
-
-  @Preconditions
-    UART1_Initialize() function should have been called
-    for the ISR to execute correctly.
-
-  @Param
-    None
-
-  @Returns
-    None
-*/
-void UART1_RxDataHandler(void);
 
 /**
   @Summary
@@ -499,79 +433,11 @@ void UART1_SetErrorHandler(void (* interruptHandler)(void));
 
 
 
-/**
-  @Summary
-    UART1 Receive Interrupt Handler
-
-  @Description
-    This is a pointer to the function that will be called upon UART1 receive interrupt
-
-  @Preconditions
-    Initialize  the UART1 module with receive interrupt enabled
-
-  @Param
-    None
-
-  @Returns
-    None
-*/
-void (*UART1_RxInterruptHandler)(void);
-
-/**
-  @Summary
-    UART1 Transmit Interrupt Handler
-
-  @Description
-    This is a pointer to the function that will be called upon UART1 transmit interrupt
-
-  @Preconditions
-    Initialize  the UART1 module with transmit interrupt enabled
-
-  @Param
-    None
-
-  @Returns
-    None
-*/
-void (*UART1_TxInterruptHandler)(void);
 
 
 
-/**
-  @Summary
-    Set UART1 Receive Interrupt Handler
 
-  @Description
-    This API sets the function to be called upon UART1 receive interrupt
 
-  @Preconditions
-    Initialize  the UART1 module with receive interrupt enabled before calling this API
-
-  @Param
-    Address of function to be set as receive interrupt handler
-
-  @Returns
-    None
-*/
-void UART1_SetRxInterruptHandler(void (* InterruptHandler)(void));
-
-/**
-  @Summary
-    Set UART1 Transmit Interrupt Handler
-
-  @Description
-    This API sets the function to be called upon UART1 transmit interrupt
-
-  @Preconditions
-    Initialize  the UART1 module with transmit interrupt enabled before calling this API
-
-  @Param
-    Address of function to be set as transmit interrupt handler
-
-  @Returns
-    None
-*/
-void UART1_SetTxInterruptHandler(void (* InterruptHandler)(void));