+++ examples/noIP-PIC18-Curiosity-Nano/PIC18.X/tc6-stub.c
@@ -50,7 +50,6 @@ Microchip or any third party.
 #include "tc6-conf.h"
 #include "tc6.h"
 #include "systick.h"
-#include "DMASPItransfer.h"/*THW220107 add include file */
 
 /*>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>*/
 /*                          USER ADJUSTABLE                             */
@@ -64,6 +63,7 @@ Microchip or any third party.
 #define ASSERT(x)
 #endif
 
+#define TC6_CHIP_SELECT_INSTANCE      (0)
 static const uint8_t FALLBACK_MAC[] = {0x00u, 0x80u, 0xC2u, 0x00u, 0x01u, 0xCCu};
 
 /*>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>*/
@@ -73,13 +73,10 @@ static const uint8_t FALLBACK_MAC[] = {0x00u, 0x80u, 0xC2u, 0x00u, 0x01u, 0xCCu}
 typedef struct
 {
     uint8_t mac[6];
-    uint8_t intIn;
-    uint8_t intOut;
-    uint8_t intReported;
     uint8_t idx;
     bool opened;
-    bool busy;
     bool firstTransaction;
+    volatile bool busy;
 } Stub_Local_t;
 
 static Stub_Local_t d[TC6_MAX_INSTANCES] = { 0 };
@@ -106,33 +103,13 @@ bool TC6Stub_Init(uint8_t idx, uint8_t pMac[6])
         memcpy(pMac, FALLBACK_MAC, 5u);
         pMac[5] = idx;
 
-        IO_RST1_SetLow();
-        SysTick_DelayMS(10);
-        IO_RST1_SetHigh();
+        IO_RST3_SetHigh();
         SysTick_DelayMS(10);
 
         INT0_SetInterruptHandler(IntHandler);
-        SPI1_Open(SPI1_DEFAULT); /* open SPI instance */
-        vidInitDmaSpiTransfer();
-        DMA2_SetDCNTIInterruptHandler(&vidDma2DestCountISR);
         success = true;
     }
     return success;
-}
-
-bool TC6Stub_IntActive(uint8_t idx)
-{
-    Stub_Local_t *ps = &d[idx];
-    ASSERT(idx < TC6_MAX_INSTANCES);
-    ps->intReported = ps->intIn;
-    return (ps->intReported != ps->intOut);
-}
-
-void TC6Stub_ReleaseInt(uint8_t idx)
-{
-    Stub_Local_t *ps = &d[idx];
-    ASSERT(idx < TC6_MAX_INSTANCES);
-    ps->intOut = ps->intReported;
 }
 
 uint32_t TC6Stub_GetTick(void)
@@ -140,53 +117,31 @@ uint32_t TC6Stub_GetTick(void)
     return SysTick_GetMillis();
 }
 
-bool TC6Stub_SpiTransaction(uint8_t idx, uint8_t *pTx, uint8_t *pRx, uint16_t len)
+bool TC6Stub_SpiTransaction(uint8_t idx, const uint8_t *pTx, uint8_t *pRx, uint16_t len)
 {
-    uint16_t i;
-    if (idx > FIRST_TC6_INSTANCE) {
-        TC6_ASSERT(false);
-        return false;
-    }
-    /*
-     * provide information to DMA to transfer data via DMA <-> SPI
-     */
-    (void)u8SpiStartTransaction(pTx, pRx, len);
-    /* wait until transfer finished */
-    while (DMA_TRANSFER_PROGRESS == u8GetProcessStatus());
+    IO_CS3_SetLow();
+    SPI1_ExchangeBlocks(pTx, pRx, len);
+    IO_CS3_SetHigh();
     return true;
+}
+
+void TC6Stub_IntPinInterruptEnable(uint8_t idx, bool enableInt)
+{
+    if(enableInt) {
+        INTERRUPT_GlobalInterruptEnable();
+    } else {
+        INTERRUPT_GlobalInterruptDisable();
+    }
 }
 
 /*>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>*/
 /*                  PRIVATE FUNCTION IMPLEMENTATIONS                    */
 /*>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>*/
 
-
 static void IntHandler(void)
 {
     Stub_Local_t *ps = &d[0];
-    ps->intIn++;
+    TC6_HandleMacPhyInterrupt(ps->idx);
     EXT_INT0_InterruptFlagClear();
 }
 
-/*>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>*/
-/*                     CALLBACK FROM DMA SPI DRIVER                     */
-/*>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>*/
-
-void vidChipSelectHigh(void)
-{
-    IO_CS1_SetHigh();
-}
-
-void vidChipSelectLow(void)
-{
-    /* SPI Driver is sending corrupted data on the first transaction, silently drop it by not toggling the CS */
-    if (!d[0].firstTransaction) {
-        IO_CS1_SetLow();
-    }
-}
-
-void vidTransactionDone(void)
-{
-    d[0].firstTransaction = false;
-    TC6_SpiBufferDone(0, true);
-}